organized components folders and cleaned up index file
